Split code into components Update layout Add media queries for mobile responsiveness Update button styles
