--- src/App.css
@@ -1,75 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=B612+Mono|Lato&display=swap');
-.App {
-  /* margin: 0 auto; */
-  text-align: center;
-  background-color: #ffffff;
-  /* padding-top:60px; */
-  background-image: url('./images/speech_to_text.svg');
-  background-repeat: no-repeat;
-  background-size: contain;
-  font-family: 'Lato', sans-serif;
-}
-.container{
-  display: flex;
-  justify-content: flex-end;
-}
-.content{
-  width:60vw;
-}
-.title{
-  color:#453adf;
-  font-size: 65px;
-  font-family: 'B612 Mono', monospace;
-}
-
-.subTitle{
-  font-weight: 300;
-  color:#6c61f6;
-}
-
-.start, .stop {
-  padding:10px;
-  background-color: #6c61f6;
-  color:#ffffff;
-  font-weight: bold;
-  font-size: 14px;
-  border-radius:5px;
-}
-
-.textarea{
-  overflow: auto;
-  width: 500px;
-  height: 300px;
-  margin: 20px 20px;
-  border:none;
-  border-radius:10px;
-  box-shadow: 0 0 3px 5px rgba(0,0,0,0.12);
-  padding:15px;
-  font-size:14px;
-}
-
-.btn{
-  padding:10px;
-  background-color: #39364d;
-  color:#ffffff;
-  font-size:14px;
-  border-radius:5px;
-  font-weight:bold;
-  margin:0 5px 20px 5px;
-}
-
-.pulse {
-  transition: all 200ms ease;
-  animation: pulse 2s infinite ease-in-out;
-}
-.footer{
-  padding:20px;
-  color: #ffffff;
-  background-color: #4c4c56;
-  font-weight: 300;
-}
-@keyframes pulse {
-  0%    {opacity: 1;}
-  50%   {opacity: .3;}
-  100%  {opacity: 1;}
-}
