+++ src/components/App.css
@@ -1,0 +1,63 @@
+@import url('https://fonts.googleapis.com/css?family=B612+Mono|Lato&display=swap');
+.App {
+  /* margin: 0 auto; */
+  text-align: center;
+  background-color: #ffffff;
+  font-family: 'Lato', sans-serif;
+}
+.container{
+  display: flex;
+  padding: 10px 15px;
+}
+.col-1{
+  background-image: url('../images/speech_to_text.svg');
+  flex:1;
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+.col-2{
+  flex:1;
+}
+.subTitle{
+  font-weight: 300;
+  color:#6c61f6;
+}
+
+.textarea{
+  overflow: auto;
+  width: 80%;
+  height: 300px;
+  margin: 20px 20px;
+  border:none;
+  border-radius:10px;
+  box-shadow: 0 0 3px 5px rgba(0,0,0,0.12);
+  padding:15px;
+  font-size:14px;
+}
+
+.pulse {
+  transition: all 200ms ease;
+  animation: pulse 2s infinite ease-in-out;
+}
+
+@keyframes pulse {
+  0%    {opacity: 1;}
+  50%   {opacity: .3;}
+  100%  {opacity: 1;}
+}
+
+@media (min-width: 320px) and (max-width: 480px) {
+  
+  .col-1{
+    display: none;
+  }
+  .container {
+    flex-direction: column;
+  }
+  .title{
+    font-size: 45px !important;
+  }
+  .textarea{
+    height: 246px;
+  }
+}
