+++ src/components/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Footer.css';
+
+function Footer (){
+    return(
+        <footer className="footer">
+        <p>Copyright &copy; {new Date().getFullYear()} </p>
+        </footer>
+    )
+}
+
+
+export default Footer;
