+++ src/components/Header.css
@@ -1,0 +1,6 @@
+.title{
+    color:#453adf;
+    font-size: 65px;
+    font-family: 'B612 Mono', monospace;
+    margin:8px 0;
+  }
