+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from "react";
+import './Header.css';
+
+function Header(){
+    return(
+        <header>
+        <h1 className="title">Vext</h1>
+        </header>
+    )
+}
+
+
+
+export default Header;
